+++ Cargo.toml
@@ -3,7 +3,7 @@ name = "elastic"
 version = "0.3.0"
 authors = ["Ashley Mannix <ashleymannix@live.com.au>"]
 license = "Apache-2.0"
-description = "A client for the Elassticsearch REST API."
+description = "A client for the Elasticsearch REST API."
 documentation = "https://docs.rs/elastic/"
 repository = "https://github.com/elastic-rs/elastic"
